+++ src/main/java/com/cgm/infolab/db/repository/queryhelper/UserQueryHelper.java
@@ -1,7 +1,6 @@
-package com.cgm.infolab.db.repository;
+package com.cgm.infolab.db.repository.queryhelper;
 
 import com.cgm.infolab.db.model.Username;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
