+++ src/main/java/com/cgm/infolab/db/repository/queryhelper/UserQueryResult.java
@@ -1,8 +1,7 @@
-package com.cgm.infolab.db.repository;
+package com.cgm.infolab.db.repository.queryhelper;
 
 import com.cgm.infolab.db.model.Username;
 import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
